Converted JS files into modules.
