+++ app/static/js/helpers.mjs
@@ -1,0 +1,44 @@
+function deleteChildrenById(id){
+    const node = document.getElementById(id);
+    while (node.firstChild) {
+        node.removeChild(node.firstChild);
+  }
+}
+
+function hideById(id){
+  let node = document.getElementById(id);
+  node.classList.remove('visible');
+  node.classList.add('hidden');
+}
+
+function showById(id){
+  let node = document.getElementById(id);
+  node.classList.add('visible');
+  node.classList.remove('hidden');
+}
+
+function deleteChildren(node){
+  while (node.firstChild) {
+      node.removeChild(node.firstChild);
+    }
+}
+
+async function postData(url = '', data = {}) {
+  // Default options are marked with *
+  const response = await fetch(url, {
+    method: 'POST', // *GET, POST, PUT, DELETE, etc.
+    mode: 'cors', // no-cors, *cors, same-origin
+    cache: 'no-cache', // *default, no-cache, reload, force-cache, only-if-cached
+    credentials: 'same-origin', // include, *same-origin, omit
+    headers: {
+      'Content-Type': 'application/json'
+      // 'Content-Type': 'application/x-www-form-urlencoded',
+    },
+    redirect: 'follow', // manual, *follow, error
+    referrer: 'no-referrer', // no-referrer, *client
+    body: JSON.stringify(data) // body data type must match "Content-Type" header
+  });
+  return await response.json(); // parses JSON response into native JavaScript objects
+}
+
+export { deleteChildrenById, hideById, showById, deleteChildren, postData };
