+++ app/templates/analysis.html
@@ -32,10 +32,9 @@
         PRELOADED_MAP = {{ preloaded_map|tojson|safe}};
         {% endif %}
     </script>
-<script src="{{ url_for('static',filename='js/us_state_geojson.js') }}"></script>
-<script src="{{ url_for('static',filename='js/analysis.js') }}"></script>
+
 <script src="{{ url_for('static',filename='js/papaparse.min.js') }}"></script>
-<script src="{{ url_for('static',filename='js/upload_data.js') }}"></script>
+<script type="module" src="{{ url_for('static',filename='js/main.js') }}"></script>
 
 
 {% endblock %}
