+++ app/templates/base.html
@@ -5,7 +5,7 @@
     <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='styles.css') }}">
     <link rel="shortcut icon" href="{{ url_for('static', filename='favicon.ico') }}">
     <script type="text/javascript">
-        const SCRIPT_ROOT = "{{ url_for('index', _external=True) }}";
+        window.SCRIPT_ROOT = "{{ url_for('index', _external=True) }}";
     </script>
     <meta charset="utf-8">
     <title>American Health Mapper</title>
